Revert "x"
This reverts commit 7e40d67e9f85d1752272ad956176179f39ce5da4.

+++ cores/arduino/zephyrCommon.cpp
@@ -315,7 +315,7 @@ void tone(pin_size_t pinNumber, unsigned int frequency,
 
   key = k_spin_lock(&pt->lock);
   pt->infinity = (duration == 0);
-  pt->count = min((uint64_t)duration * frequency * TOGGLES_PER_CYCLE / MSEC_PER_SEC, UINT32_MAX);
+  pt->count = min((uint64_t)duration * frequency * TOGGLES_PER_CYCLE / MSEC_PER_SEC, (uint64_t)UINT32_MAX);
   pt->pin = pinNumber;
   k_spin_unlock(&pt->lock, key);
